requirements.txt
asgiref==3.7.2
Django==4.2.8
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.4.1
haversine==2.8.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0
